+++ front/lib/app/config/routes.dart
@@ -7,9 +7,9 @@ import 'package:front/app/domain/presentation/main/screen/main_screen.dart';
 import 'package:front/app/domain/presentation/project/screen/project_creation.dart';
 import 'package:front/app/domain/presentation/project/screen/project_detail.dart';
 import 'package:front/app/domain/presentation/project/screen/project_update.dart';
-import 'package:front/presentation/pages/team/team_create.dart';
-import 'package:front/presentation/pages/team/team_datail.dart';
-import 'package:front/presentation/pages/team/teams_list.dart';
+import 'package:front/features/team/presentation/pages/team/team_create.dart';
+import 'package:front/features/team/presentation/pages/team/team_datail.dart';
+import 'package:front/features/team/presentation/pages/team/teams_list.dart';
 import 'package:go_router/go_router.dart';
 
 final router = GoRouter(initialLocation: '/', routes: [
