+++ front/lib/app/domain/presentation/project/component/project_manager_selector.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:front/core/user/models/user.dart';
-import 'package:front/presentation/pages/team/widgets/dialog.dart';
+import 'package:front/features/team/presentation/pages/team/widgets/dialog.dart';
 
 typedef VoidCallback = void Function(UserModel user);
