+++ front/lib/app/locator.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:front/app/di/team_di.dart';
-import 'package:front/features/data/data_source/team_remote_data_source.dart';
+import 'package:front/features/team/data/data_source/team_remote_data_source.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
 import 'package:front/features/team/usecases/answer_to_invitation_usecase.dart';
 import 'package:front/features/team/usecases/create_team_usecase.dart';
