+++ front/lib/core/utils/api_response.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:front/features/data/models/team.dart';
+import 'package:front/features/team/data/models/team.dart';
 
 part 'api_response.g.dart';
