+++ front/lib/features/team/data/api/team_api.dart
@@ -1,8 +1,8 @@
 import 'package:dio/dio.dart' hide Headers;
-import 'package:front/features/data/models/invite_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
-import 'package:front/features/data/models/team.dart';
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/team.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:retrofit/http.dart';
