+++ front/lib/features/team/data/data_source/team_remote_data_source.dart
@@ -1,7 +1,7 @@
-import 'package:front/features/data/models/invite_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
-import 'package:front/features/data/models/team.dart';
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/team.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/core/utils/api_response.dart';
 
 abstract class TeamRemoteDataSource {
