+++ front/lib/features/team/data/data_source/team_remote_data_source_impl.dart
@@ -1,9 +1,9 @@
 import 'package:front/core/network_handling/app_dio.dart';
-import 'package:front/features/data/api/team_api.dart';
-import 'package:front/features/data/models/invite_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
-import 'package:front/features/data/models/team.dart';
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/api/team_api.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/team.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/core/utils/api_response.dart';
 
 import 'team_remote_data_source.dart';
