+++ front/lib/features/team/data/models/team_detail.dart
@@ -1,5 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:front/features/data/models/team_member.dart';
+import 'package:front/features/team/data/models/team_member.dart';
 
 part 'team_detail.g.dart';
