+++ front/lib/features/team/entities/team_entity.dart
