+++ front/lib/features/team/presentation/pages/team/team_datail.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:front/features/data/models/team_detail.dart';
-import 'package:front/features/team/state/projects_state.dart';
-import 'package:front/features/team/viewmodel/team_detail.dart';
-import 'package:front/presentation/pages/team/widgets/dialog.dart';
-import 'package:front/presentation/pages/team/widgets/selecting_sharing_method_dialog.dart';
-import 'package:front/presentation/providers/team_controller.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
+import 'package:front/features/team/presentation/pages/team/widgets/dialog.dart';
+import 'package:front/features/team/presentation/pages/team/widgets/selecting_sharing_method_dialog.dart';
+import 'package:front/features/team/presentation/providers/projects_state.dart';
+import 'package:front/features/team/presentation/providers/team_controller.dart';
+import 'package:front/features/team/presentation/providers/team_detail.dart';
 
 class TeamDetailScreen extends ConsumerStatefulWidget {
   final String teamId;
