+++ front/lib/features/team/presentation/pages/team/widgets/dialog.dart
