+++ front/lib/features/team/presentation/pages/team/widgets/share_via_qr_dialog.dart
