+++ front/lib/features/team/presentation/pages/team/widgets/team_member_clips.dart
