+++ front/lib/features/team/presentation/providers/team_detail.dart
@@ -1,5 +1,5 @@
 import 'package:front/core/project/domain/usecases/riverpod.dart';
-import 'package:front/features/team/state/projects_state.dart';
+import 'package:front/features/team/presentation/providers/projects_state.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'team_detail.g.dart';
