+++ front/lib/features/team/repositories/team_repository.dart
@@ -1,7 +1,7 @@
-import 'package:front/features/data/models/invite_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
-import 'package:front/features/data/models/team.dart';
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/team.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:front/core/utils/result.dart';
