+++ front/lib/features/team/repositories/team_repository_impl.dart
@@ -1,8 +1,8 @@
-import 'package:front/features/data/data_source/team_remote_data_source.dart';
-import 'package:front/features/data/models/invite_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
-import 'package:front/features/data/models/team.dart';
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/data_source/team_remote_data_source.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/team.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:front/core/utils/result.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
