+++ front/lib/features/team/usecases/answer_to_invitation_usecase.dart
@@ -1,4 +1,4 @@
-import 'package:front/features/data/models/invite_response.dart';
+import 'package:front/features/team/data/models/invite_response.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:front/core/utils/result.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
