+++ front/lib/features/team/usecases/get_team_by_id_usecase.dart
@@ -1,4 +1,4 @@
-import 'package:front/features/data/models/team_detail.dart';
+import 'package:front/features/team/data/models/team_detail.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:front/core/utils/result.dart';
