+++ front/lib/features/team/usecases/get_teams_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:front/features/data/models/team.dart';
+import 'package:front/features/team/data/models/team.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
 import 'package:front/core/utils/api_response.dart';
 import 'package:front/core/utils/result.dart';
