+++ front/lib/features/team/usecases/invite_team_by_email_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:front/core/utils/api_response.dart';
-import 'package:front/features/data/models/invite_team.dart';
+import 'package:front/features/team/data/models/invite_team.dart';
 import 'package:front/features/team/repositories/team_repository.dart';
 import 'package:front/core/utils/result.dart';
